Trying to follow guide on refactoring
but really i have no idea what's going on
Moved all the sub-views as internal structs for gallery and editor
