+++ LetUsCook/Model/AppController.swift
@@ -1,0 +1,17 @@
+//
+//  AppController.swift
+//  LetUsCook
+//
+//  Created by Justin Hoang on 4/14/24.
+//
+
+import Foundation
+
+@Observable
+class AppController {
+    var dataModel: DataModel
+    
+    init() {
+        dataModel = DataModel()
+    }
+}
