+++ LetUsCook/View/MainView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct MainView: View {
     @Environment(\.modelContext) private var modelContext
-
+    
     @State private var sidebarSelection: SidebarItem? = .Gallery
     @State private var recipeSelection: Recipe?
 
@@ -27,9 +27,4 @@ struct MainView: View {
             RecipeView(recipe: recipeSelection)
         }
     }
-}
-
-#Preview {
-    MainView()
-        .modelContainer(previewContainer)
 }
